requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
cachetools==4.2.2
certifi==2021.10.8
dnspython==2.2.1
pycodestyle==2.9.1
pymongo==4.0.1
python-dotenv==0.21.0
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
